+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [//]: # "## [unreleased] - yyyy-mm-dd"
 
 
-## [unreleased] - yyyy-mm-dd
+## [0.4.1] - 2025-11-17
 
 ### Added